fix(ci): resolve clippy warnings and cross-platform test failures
- utils.rs: mark Linux-only desktop detection helpers #[allow(dead_code)]
- word_builder.rs: drop useless .into_iter() in zip
- timing.rs: drop tautological .min(usize::MAX) on 64-bit
- boundaries.rs: drop redundant 'snapped as i64' cast (already i64)
- transcription_mock: derive Default, drop .trim() before split_whitespace
- export test: use tempdir/nonexistent subdir so invalid-path test fails
  on Linux too (Windows drive-letter path is a valid filename on Linux)
- .gitignore: un-ignore features/*/fixtures/ so project_v1_0_0 fixture
  is tracked; project.rs test now loads the committed fixture in CI

+++ src-tauri/src/managers/splice/boundaries.rs
@@ -362,7 +362,7 @@ mod tests {
         let snapped = snap_to_zero_crossing(2_500, &buf, sr, DEFAULT_SNAP_RADIUS_US);
         // Convert back to sample index; integer truncation can land us one
         // sample before the true snap point, so accept a ±1 window.
-        let sample = ((snapped as i64 * sr as i64) / 1_000_000) as usize;
+        let sample = ((snapped * sr as i64) / 1_000_000) as usize;
         let window_lo = sample.saturating_sub(1);
         let window_hi = (sample + 2).min(buf.len() - 1);
         let mut found = false;
